crm/requirements.txt
flask>=3.0.0
bcrypt>=4.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
markdown>=3.5
jinja2>=3.1.0
gunicorn>=21.2.0
